+++ LiVHY/src/objects/object.cpp
@@ -16,7 +16,7 @@ Object::Object(const Material* mat, bool cull, bool samplingTarget)
 }
 
 
-bool Object::Hit(const Ray& ray, float tMin, float tMax, HitInfo& hitInfo) const
+bool Object::Hit(const Ray& ray, double tMin, double tMax, HitInfo& hitInfo) const
 {
 	return true;
 }
@@ -28,12 +28,12 @@ bool Object::BoundingBox(AABB& outputBox) const
 
 Vec3 Object::Random(Vec3 point) const
 {
-	return Vec3(1.0f, 0.0f, 0.0f);
+	return Vec3(1, 0, 0);
 }
 
-float Object::PdfValue(const Vec3& origin, const Vec3& dir) const
+double Object::PdfValue(const Vec3& origin, const Vec3& dir) const
 {
-	return 0.0f;
+	return 0.0;
 }
 
 bool Object::IsSamplingTarget() const
